Modularize semantic analyzer shell
