+++ analyzer/semantic/index.ts
@@ -1,0 +1,2 @@
+export * from './SemanticAnalyzer.js';
+export * from './types.js';
